[W.I.P.] Add HLMDLv10 parser
This commit only contains a part of the HLMDLv10 parser. It is able to
load and read most of the format (not everything). Keep in mind that
the code is not perfect as some hacks are present and some stuff could
be handled/reworked in a better way.

Reminder: this is the **parser**, nothing about the rendering part has
been changed. As of right now, TrenchBroom is only capable of
rendering "key frames models" (Quake, Quake II, Quake III) and not
"skeletal models". There is also no concept of "bodygroup".

The unit test code is incomplete, also test assets were not included.